requirements.txt
fastapi
pydantic
python-dotenv
redis
requests
uvicorn
python-multipart
